+++ modules/ui/fields/index.js
@@ -3,7 +3,7 @@ export * from './combo';
 export * from './input';
 export * from './access';
 export * from './address';
-export * from './cycleway';
+export * from './directional_combo';
 export * from './lanes';
 export * from './localized';
 export * from './roadheight';
@@ -46,7 +46,7 @@ import {
 
 import { uiFieldAccess } from './access';
 import { uiFieldAddress } from './address';
-import { uiFieldCycleway } from './cycleway';
+import { uiFieldDirectionalCombo } from './directional_combo';
 import { uiFieldLanes } from './lanes';
 import { uiFieldLocalized } from './localized';
 import { uiFieldRoadheight } from './roadheight';
@@ -62,8 +62,9 @@ export var uiFields = {
     check: uiFieldCheck,
     colour: uiFieldColour,
     combo: uiFieldCombo,
-    cycleway: uiFieldCycleway,
+    cycleway: uiFieldDirectionalCombo,
     defaultCheck: uiFieldDefaultCheck,
+    directionalCombo: uiFieldDirectionalCombo,
     email: uiFieldEmail,
     identifier: uiFieldIdentifier,
     lanes: uiFieldLanes,